Fix issue #30
Before this change, comments starting with --, like it is the case in
Haskell would confuse github-review if they are deleted. After this
change it is no longer the case.
Concretely, the parsing logic changes to:
- Only look at --- and +++ when they happen at the top level of a
review (pos is not net)
- Reset the pos to nil when we hit a line that starts with `diff` to
indicate that we are at the top level again

- [X] Took a partial diff from hledger a haskell project and added comments
in multiple parts of it with multiple files and multiple comments
deleted. Added a unit test for it.

FYI @symbiont-sam-halliday who reported the issue and
@matthew-piziak who reacted on the issue