Introduce major mode for code review
Before this change we were using the diff mode as major mode for code
review. After this change we use the github-review-mode. For now the
diff-mode and github-review-mode behave the same (we use derived mode)
but in the future we will add new features to the github-review-mode
specific to code review.

- [x] Launched a code review and made sure that the mode is set properly and
diff mode navigation still works fine
- [ ] CI